Make ArchGDAL a weak dependency via extension
- Move _reproject_ibcao_to_netcdf into NumericalEarthArchGDALExt so ArchGDAL
  is only required when loading IBCAO data, not at package load time
- ArchGDAL moved from [deps] to [weakdeps] in Project.toml
- ArchGDAL added as regular dep in test/Project.toml
