+++ Project.toml
@@ -6,7 +6,6 @@ authors = ["NumericalEarth contributors"]
 
 [deps]
 Adapt = "79e6a3ab-5dfb-504d-930d-738a2a938a0e"
-ArchGDAL = "c9ce4bd3-c3d5-55b8-8973-c0e20141b8c3"
 CFTime = "179af706-886a-5703-950a-314cd64e0468"
 ClimaSeaIce = "6ba0ff68-24e6-4315-936c-2e99227c95a4"
 DataDeps = "124859b0-ceae-595e-8997-d05f6a7a8dfe"
@@ -32,6 +31,7 @@ Thermodynamics = "b60c26fb-14c3-4610-9d3e-2d17fe7ff00c"
 ZipFile = "a5390f91-8eb1-5f08-bee0-b1d1ffed6cea"
 
 [weakdeps]
+ArchGDAL = "c9ce4bd3-c3d5-55b8-8973-c0e20141b8c3"
 Breeze = "660aa2fb-d4c8-4359-a52c-9c057bc511da"
 CDSAPI = "8a7b9de3-9c00-473e-88b4-7eccd7ef2fea"
 CondaPkg = "992eb4ea-22a4-4c89-a5bb-47a3300528ab"
@@ -43,6 +43,7 @@ WorldOceanAtlasTools = "04f20302-f1b9-11e8-29d9-7d841cb0a64a"
 XESMF = "2e0b0046-e7a1-486f-88de-807ee8ffabe5"
 
 [extensions]
+NumericalEarthArchGDALExt = "ArchGDAL"
 NumericalEarthBreezeExt = "Breeze"
 NumericalEarthCDSAPIExt = "CDSAPI"
 NumericalEarthCopernicusMarineExt = "CopernicusMarine"
@@ -54,6 +55,7 @@ NumericalEarthWOAExt = "WorldOceanAtlasTools"
 [compat]
 Adapt = "4"
 ArchGDAL = "0.10"
+
 Breeze = "0.4"
 CDSAPI = "2.2.1"
 CFTime = "0.1, 0.2"
