+++ ext/NumericalEarthArchGDALExt.jl
@@ -1,0 +1,50 @@
+module NumericalEarthArchGDALExt
+
+using NumericalEarth
+using ArchGDAL
+using NCDatasets
+
+import NumericalEarth.DataWrangling.IBCAO: _reproject_ibcao_to_netcdf
+
+function _reproject_ibcao_to_netcdf(tiff_path, nc_path)
+    ArchGDAL.read(tiff_path) do src
+        # Warp from EPSG:3996 (Polar Stereographic) to EPSG:4326 (WGS84)
+        # at 0.01° resolution, clipping to 64–90°N
+        ArchGDAL.gdalwarp([src],
+            ["-t_srs", "EPSG:4326",
+             "-te",    "-180", "64", "180", "90",  # xmin ymin xmax ymax
+             "-tr",    "0.01", "0.01",             # target resolution (degrees)
+             "-r",     "bilinear",                 # resampling method
+             "-ot",    "Float32"]) do warped
+
+            # ArchGDAL returns data as (Nx, Ny) with y from north to south (GDAL convention)
+            data = Float32.(ArchGDAL.read(warped, 1))
+            data = data[:, end:-1:1]  # flip y: j=1 → 64°N (south), j=Ny → 90°N (north)
+
+            Nx, Ny = size(data)  # expected: (36000, 2600)
+
+            NCDataset(nc_path, "c") do ds
+                defDim(ds, "lon", Nx)
+                defDim(ds, "lat", Ny)
+
+                lon_var = defVar(ds, "lon", Float64, ("lon",);
+                                attrib = ["units" => "degrees_east",
+                                          "long_name" => "longitude"])
+                lat_var = defVar(ds, "lat", Float64, ("lat",);
+                                attrib = ["units" => "degrees_north",
+                                          "long_name" => "latitude"])
+                z_var   = defVar(ds, "z",   Float32, ("lon", "lat");
+                                attrib = ["long_name" => "elevation",
+                                          "units"     => "m"])
+
+                lon_var[:] = range(-180 + 0.005, 180 - 0.005; length=Nx)
+                lat_var[:] = range(64 + 0.005, 90 - 0.005; length=Ny)
+                z_var[:, :] = data
+            end
+        end
+    end
+
+    return nothing
+end
+
+end # module NumericalEarthArchGDALExt
